(BSR)[API] refactor: Move upsert_venue_banner tests to its own module
